Updated the method to check for the result_ids

+++ Python Examples/TestMonitor/test_data_manager_client.py
@@ -156,16 +156,18 @@ def raise_post_request(uri, body):
     return request_response
 
 def delete_result(result_id, delete_steps = True):
+    if result_id == None :
+        raise ValueError("Missing required parameter 'result_id' when calling ResultsApi->DeleteResultV2")
     request_uri = f"{base_uri}{delete_result_route}/{result_id}?deleteSteps{delete_steps}"
     request_response = requests.delete(request_uri, headers=headers)
 
     request_response.raise_for_status()
 
-def delete_results(result_ids, delete_steps = True):    
+def delete_results(result_ids, delete_steps = True):
+    if len(result_ids) == 0 or result_ids == None:
+        raise ValueError("result_ids is a required property for DeleteResultsRequest and cannot be null or empty")
     request_uri = f"{base_uri}{delete_results_route}"
     body = delete_results_request(result_ids, delete_steps)   
     request_response =  requests.post(request_uri, json=body, headers=headers)
     
     request_response.raise_for_status()
-
-
